Add RequestDetail and CniRequestComponent to routeManagement.js
