+++ src/routes/routeManagement.js
@@ -19,6 +19,9 @@ import Notification from "../pages/other/notification";
 import Administration from "../pages/other/admin";
 import Transaction from "../pages/other/transaction";
 import Requests from "../pages/other/requests";
+import RequestDetail from "../pages/other/requestDetail";
+import RequestComponent from "../components/requests/requestComponent";
+import CniRequestComponent from "../components/requests/cniRequestComponent";
 import { GetToken } from "../services/tokenService";
 import Error from "../pages/error";
 
@@ -47,7 +50,21 @@ const RouteManagement = () => {
             <Route path="/" element={<IndexOther />}>
                 <Route index element={<Overview />} />
                 <Route path="/transaction" element={<Transaction />} />
-                <Route path="/requests" element={<Requests />} />
+
+
+
+                <Route path="/requests/" element={<Requests />} >
+                    <Route index element={<RequestComponent />} />
+                    <Route path=":requestId/" element={<RequestDetail />}>
+                        <Route index element={<AccountUserComponent />} />
+                        <Route
+                            path="docs"
+                            element={<CniRequestComponent />}
+                        />
+                    </Route>
+                </Route>
+                
+                
 
                 <Route path="/users/" element={<Users />}>
                     <Route index element={<UserComponent />} />
